+++ src/main/java/feminine_beauty/api/dtos/servico/DadosAtualizacaoServico.java
@@ -6,5 +6,5 @@ import jakarta.validation.constraints.NotNull;
 public record DadosAtualizacaoServico(
         @NotNull
         Long id,
-        Integer preco) {
+        Double preco) {
 }
